requirements.txt
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
columns==0.0.2
cors==1.0.1
et-xmlfile==1.1.0
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.7
future==1.0.0
gevent==24.2.1
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-aiplatform==1.42.1
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.2
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
googletrans==4.0.0rc1
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imageio==2.34.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy_loader==0.3
libretranslatepy==2.1.1
llvmlite==0.42.0
lxml==5.2.1
MarkupSafe==2.1.5
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
onnxruntime==1.17.1
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMatting==1.1.12
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytweening==1.2.0
pytz==2024.1
referencing==0.34.0
rembg==2.0.56
requests==2.31.0
requests-file==2.0.0
rfc3986==1.5.0
rpds-py==0.18.0
rsa==4.9
rubicon-objc==0.4.7
scikit-image==0.22.0
scipy==1.12.0
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
sympy==1.12
tifffile==2024.2.12
tldextract==5.1.2
tqdm==4.66.2
translate==3.6.1
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
zope.event==5.0
zope.interface==6.3
